--- app/src/scripts/API/APITypes/APIErrorResponse.ts
@@ -1,3 +1,0 @@
-export type APIErrorResponse = {
-    [key: string]: string;
-}
