--- app/src/scripts/API/APITypes/ConfirmationMessage.ts
@@ -1,3 +1,0 @@
-export type ConfirmationMessage = {
-    message: string;
-}
