+++ app/src/scripts/API/APITypes/Tokens.ts
@@ -1,3 +1,4 @@
+/* Simple pair of tokens. */
 export type TokenPair = {
     access_token: string;
     refresh_token: string;
