+++ app/src/scripts/API/Responses/ErrorResponse.ts
@@ -1,5 +1,9 @@
 import APIResponse from "./APIResponse.ts";
 
+/**
+ * This class inherits from APIResponse and represents an unsuccessful response from the API.
+ * It can be used to diagnose the reason for the request's failure.
+ */
 export default class ErrorResponse implements APIResponse<object> {
 
     public readonly _errorCode  : number;
