+++ app/src/scripts/Models/UserModel.ts
@@ -1,3 +1,6 @@
+/**
+ * A classic model that provides handling methods for actions that concerns a User.
+ */
 export default abstract class UserModel{
     constructor() {}
 }
